% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clas_exp}
\alias{clas_exp}
\title{An example of a classification model explainer}
\format{
A list containing the model and related information.
}
\usage{
data(clas_exp)
}
\description{
An example of a model explainer (created using the DALEX package) for a classification problem.
}
\keyword{datasets}
