% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_breakdown.R
\name{map_breakdown}
\alias{map_breakdown}
\title{Predict break down}
\usage{
map_breakdown(
  explainer,
  raster_obs,
  maxcell = 1000,
  ...,
  N = NULL,
  type = "break_down"
)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the \code{\link[DALEX:explain]{DALEX::explain()}} function}

\item{raster_obs}{a raster object with the observations to be explained (predictors used in the model)}

\item{maxcell}{the maximum number of cells in the raster. If the number of cells in the raster is greater than \code{maxcell}, the function will sample \code{maxcell} cells from the raster. By default 1000}

\item{...}{other parameters that will be passed to \code{\link[iBreakDown:break_down]{iBreakDown::break_down()}}}

\item{N}{the maximum number of observations used for calculation of attributions. By default NULL (use all)}

\item{type}{the type of variable attributions. Either \code{break_down} or \code{break_down_interactions}}
}
\value{
A raster object with the same dimensions as \code{raster_obs}. The number of layers equal to the number of variables in the model plus one.
}
\description{
This function calculates the attributions of the model
(Break Down explanations) for each observation in the raster.
Model predictions are decomposed into parts that are attributed for particular variables,
and then the attributions are assigned to the corresponding cells in the raster.
See the documentation of the \code{\link[DALEX:predict_parts]{DALEX::predict_parts()}} function for more details.
}
\examples{
library(terra)
predictors_agg = rast(system.file("raster/predictors_agg.tif", package = "spatialexplain"))
data("regr_exp", package = "spatialexplain")
regr_psp_bd = map_breakdown(regr_exp, predictors_agg, maxcell = 100)
plot(regr_psp_bd)
}
\references{
Lundberg, S. (2017). A unified approach to interpreting model predictions. arXiv preprint arXiv:1705.07874.

Staniak, M., & Biecek, P. (2018). Explanations of model predictions with live and breakDown packages. arXiv preprint arXiv:1804.01955.

Biecek, P., & Burzykowski, T. (2021). Explanatory model analysis: explore, explain, and examine predictive models. Chapman and Hall/CRC.
}
