% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_oscillations.R
\name{map_oscillations}
\alias{map_oscillations}
\title{Predict oscillations}
\usage{
map_oscillations(
  explainer,
  raster_obs,
  maxcell = 1000,
  ...,
  N = 500,
  type = "oscillations_uni"
)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the \code{\link[DALEX:explain]{DALEX::explain()}} function}

\item{raster_obs}{a SpatRaster object with the observations to be explained (predictors used in the model)}

\item{maxcell}{the maximum number of cells in the raster. If the number of cells in the raster is greater than \code{maxcell}, the function will sample \code{maxcell} cells from the raster. By default 1000}

\item{...}{other parameters that will be passed to \code{\link[iBreakDown:break_down]{iBreakDown::break_down()}}}

\item{N}{the maximum number of observations used for calculation of attributions. By default 500}

\item{type}{the type of variable attributions. Either \code{oscillations_uni} (default), \code{oscillations_emp}, \code{oscillations}}
}
\value{
A SpatRaster object with the same dimensions as \code{raster_obs}. The number of layers in it is equal to the number of variables in the model.
}
\description{
This function calculates the attributions of the model
(Oscillations explanations) for each observation in the raster.
Model predictions are decomposed into parts that are attributed for particular variables,
and then the attributions are assigned to the corresponding cells in the raster.
See the documentation of the \code{\link[DALEX:predict_parts]{DALEX::predict_parts()}} function for more details.
}
\examples{
library(terra)
predictors_agg = rast(system.file("raster/predictors_agg.tif", package = "spatialexplain"))
data("regr_exp", package = "spatialexplain")
regr_psp_bd = map_oscillations(regr_exp, predictors_agg, maxcell = 100)
plot(regr_psp_bd)
}
\references{
Lundberg, S. (2017). A unified approach to interpreting model predictions. arXiv preprint arXiv:1705.07874.

Staniak, M., & Biecek, P. (2018). Explanations of model predictions with live and breakDown packages. arXiv preprint arXiv:1804.01955.

Biecek, P., & Burzykowski, T. (2021). Explanatory model analysis: explore, explain, and examine predictive models. Chapman and Hall/CRC.
}
