% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regr_exp}
\alias{regr_exp}
\title{An example of a regression model explainer}
\format{
A list containing the model and related information.
}
\usage{
data(regr_exp)
}
\description{
An example of a model explainer (created using the DALEX package) for a regression problem.
}
\keyword{datasets}
