% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialexplain-package.R
\docType{package}
\name{spatialexplain-package}
\alias{spatialexplain}
\alias{spatialexplain-package}
\title{spatialexplain: Explore and Explain Spatial Machine Learning Models}
\description{
Provides model agnostic tools for exploring and explaining spatial machine learning models. It is based on the 'DALEX' package and uses the 'terra' package for spatial raster data handling.
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
