% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spatial_parts.R
\name{predict_spatial_parts}
\alias{predict_spatial_parts}
\title{Title}
\usage{
predict_spatial_parts(
  explainer,
  raster_obs,
  maxcell = 1000,
  ...,
  N = if (substr(type, 1, 4) == "osci") 500 else NULL,
  type = "break_down"
)
}
\arguments{
\item{type}{}
}
\description{
Title
}
